requirements.txt
Flask==2.0.1
PyMuPDF==1.18.15
pytesseract==0.3.8
Pillow==8.2.0
opencv-python-headless==4.5.2.52
numpy==1.21.0
matplotlib==3.4.2
